alles hinzugefügt und irgendwie startet es erstmal

+++ stringlist.cpp
@@ -112,7 +112,7 @@ void createStringListFromDir(StringList * list, const char *dirname)
 	int count = 0;
 	int size = 0;
 	DIR *dirinfo = opendir(dirname);
-	int dirnamelen = strlen(dirname);
+	size_t dirnamelen = strlen(dirname);
 
 	if (dirinfo)
 	{
